refactor(acerca): modularize sections
