+++ src/features/acerca/Proposito.jsx
@@ -1,0 +1,17 @@
+import React from 'react';
+import styles from '../../styles/acerca.module.css';
+
+const Proposito = () => (
+  <section className={styles.section}>
+    <div className={styles.card}>
+      <h2>Propósito</h2>
+      <p>
+        Nuestra misión es facilitar el monitoreo continuo de sistemas de agua potable en zonas rurales,
+        proporcionando datos confiables para la toma de decisiones informadas que garanticen la continuidad
+        del servicio y la calidad del agua para las comunidades.
+      </p>
+    </div>
+  </section>
+);
+
+export default Proposito;
